+++ guillotina/catalog/catalog.py
@@ -27,38 +27,47 @@ global_roles_for_permission = role_permission_manager.get_roles_for_permission
 @implementer(ICatalogUtility)
 class DefaultSearchUtility(object):
 
-    async def search(self, query):
+    async def initialize(self, app):
         pass
 
-    async def get_by_uuid(self, uuid):
+    async def search(self, container, query):
         pass
 
-    async def get_object_by_uuid(self, uuid):
+    async def get_by_uuid(self, container, uuid):
         pass
 
-    async def get_by_type(self, doc_type, query={}):
+    async def get_object_by_uuid(self, container, uuid):
+        pass
+
+    async def get_by_type(self, container, doc_type, query={}):
         pass
 
     async def get_by_path(self, container, path, depth=-1, query={}, doc_type=None):
         pass
 
-    async def get_folder_contents(self, obj):
+    async def get_folder_contents(self, container, parent_uid):
         pass
 
-    async def index(self, datas):
+    async def index(self, container, datas):
         """
         {uid: <dict>}
         """
         pass
 
-    async def remove(self, uids):
+    async def update(self, container, datas):
+        """
+        {uid: <dict>}
+        """
+        pass
+
+    async def remove(self, container, uids):
         """
         list of UIDs to remove from index
         """
         pass
 
-    async def reindex_all_content(self, container):
-        """ For all Dexterity Content add a queue task that reindex the object
+    async def reindex_all_content(self, container, security=False):
+        """ For all content add a queue task that reindex the object
         """
         pass
 
@@ -72,11 +81,11 @@ class DefaultSearchUtility(object):
         """
         pass
 
-    async def get_data(self, content):
+    async def get_data(self, content, indexes=None):
         data = {}
         adapter = queryAdapter(content, ICatalogDataAdapter)
         if adapter:
-            data.update(await adapter())
+            data.update(await adapter(indexes))
         return data
 
 
@@ -119,24 +128,57 @@ class DefaultCatalogDataAdapter(object):
         value = getattr(ob, field_name, None)
         return json_compatible(value)
 
-    async def __call__(self):
+    async def load_behavior(self, behavior):
+        if IAsyncBehavior.implementedBy(behavior.__class__):
+            # providedBy not working here?
+            await behavior.load(create=False)
+
+    async def __call__(self, indexes=None):
         # For each type
-        values = {}
+        values = {
+            'type_name': self.content.type_name
+        }
 
         for schema in iter_schemata(self.content):
             behavior = schema(self.content)
-            if IAsyncBehavior.implementedBy(behavior.__class__):
-                # providedBy not working here?
-                await behavior.load(create=False)
+            loaded = False
             for index_name, index_data in merged_tagged_value_dict(schema, index.key).items():
+                if index_name in values:
+                    # you can override indexers so we do not want to index
+                    # the same value more than once
+                    continue
+
                 try:
-                    if 'accessor' in index_data:
+                    # accessors we always reindex since we can't know if updated
+                    # from the indexes param--they are "fake" like indexes, not fields
+                    if ('accessor' in index_data and (
+                            indexes is None or index_data.get('field') in indexes)):
+                        if not loaded:
+                            await self.load_behavior(behavior)
+                            loaded = True
                         values[index_name] = index_data['accessor'](behavior)
-                    else:
+                    elif (indexes is None or index_name in indexes or
+                            isinstance(getattr(type(self.content), index_name, None), property)):
+                        if not loaded:
+                            await self.load_behavior(behavior)
+                            loaded = True
+                        # in this case, properties are also dynamic so we have to make sure
+                        # to allow for them to be reindexed every time.
                         values[index_name] = self.get_data(behavior, schema, index_name)
                 except NoIndexField:
                     pass
-            for metadata_name in merged_tagged_value_list(schema, metadata.key):
-                values[metadata_name] = self.get_data(behavior, schema, metadata_name)
 
+            for metadata_name in merged_tagged_value_list(schema, metadata.key):
+                if (indexes is not None and metadata_name not in indexes and
+                        not isinstance(getattr(type(self.content), metadata_name, None), property)):
+                    # in this case, properties are also dynamic so we have to make sure
+                    # to allow for them to be reindexed every time.
+                    continue  # skip
+                if not loaded:
+                    await self.load_behavior(behavior)
+                    loaded = True
+                try:
+                    values[metadata_name] = self.get_data(behavior, schema, metadata_name)
+                except NoIndexField:
+                    pass
         return values