+++ guillotina/factory/app.py
@@ -2,6 +2,7 @@ from aiohttp import web
 from guillotina import app_settings
 from guillotina import configure
 from guillotina import cors
+from guillotina import glogging
 from guillotina import interfaces
 from guillotina import languages
 from guillotina.async import IAsyncUtility
@@ -23,15 +24,15 @@ from guillotina.interfaces import IDatabaseConfigurationFactory
 from guillotina.interfaces.content import IContentNegotiation
 from guillotina.request import Request
 from guillotina.traversal import TraversalRouter
+from guillotina.utils import lazy_apply
 from guillotina.utils import resolve_dotted_name
 from guillotina.utils import resolve_path
+from guillotina.writable import check_writable_request
 
 import aiohttp
 import asyncio
 import collections
-import inspect
 import json
-import logging
 import logging.config
 
 
@@ -41,7 +42,7 @@ except ImportError:
     RSA = None
 
 
-logger = logging.getLogger('guillotina')
+logger = glogging.getLogger('guillotina')
 
 
 def update_app_settings(settings):
@@ -56,10 +57,7 @@ def update_app_settings(settings):
 def load_application(module, root, settings):
     # includeme function
     if hasattr(module, 'includeme'):
-        args = [root]
-        if len(inspect.signature(module.includeme).parameters) == 2:
-            args.append(settings)
-        module.includeme(*args)
+        lazy_apply(module.includeme, root, settings)
     # app_settings
     if hasattr(module, 'app_settings') and app_settings != module.app_settings:
         update_app_settings(module.app_settings)
@@ -93,7 +91,8 @@ _delayed_default_settings = {
         "en-us": languages.IENUS,
         "ca": languages.ICA
     },
-    'cors_renderer': cors.DefaultCorsRenderer
+    'cors_renderer': cors.DefaultCorsRenderer,
+    'check_writable_request': check_writable_request
 }
 
 
@@ -108,7 +107,8 @@ class GuillotinaAIOHTTPApplication(web.Application):
                     label = 'TID Conflict Error detected'
                 tid = getattr(getattr(request, '_txn', None), '_tid', 'not issued')
                 logger.warning(
-                    f'{label}, retrying request, tid: {tid}, retries: {retries + 1})')
+                    f'{label}, retrying request, tid: {tid}, retries: {retries + 1})',
+                    exc_info=True)
                 request._retry_attempt = retries + 1
                 return await self._handle(request, retries + 1)
             logger.error(
@@ -123,6 +123,7 @@ class GuillotinaAIOHTTPApplication(web.Application):
             message, payload, protocol, writer, protocol._time_service, task,
             secure_proxy_ssl_header=self._secure_proxy_ssl_header,
             client_max_size=self._client_max_size)
+
 
 def make_aiohttp_application(settings, middlewares=[]):
     return GuillotinaAIOHTTPApplication(
@@ -264,7 +265,11 @@ def make_app(config_file=None, settings=None, loop=None, server_app=None):
 
     for utility in getAllUtilitiesRegisteredFor(IAsyncUtility):
         # In case there is Utilties that are registered
-        ident = asyncio.ensure_future(utility.initialize(app=server_app), loop=loop)
+        if hasattr(utility, 'initialize'):
+            ident = asyncio.ensure_future(
+                lazy_apply(utility.initialize, app=server_app), loop=loop)
+        else:
+            logger.warn(f'No initialize method found on {utility} object')
         root.add_async_task(utility, ident, {})
 
     server_app.on_cleanup.append(close_utilities)
@@ -280,7 +285,8 @@ def make_app(config_file=None, settings=None, loop=None, server_app=None):
 
 async def close_utilities(app):
     for utility in getAllUtilitiesRegisteredFor(IAsyncUtility):
-        asyncio.ensure_future(utility.finalize(app=app), loop=app.loop)
+        if hasattr(utility, 'finalize'):
+            asyncio.ensure_future(lazy_apply(utility.finalize, app=app), loop=app.loop)
     for db in app.router._root:
         if IDatabase.providedBy(db[1]):
             await db[1]._db.finalize()